Add support for linked worktrees
Within a linked worktree (ie a worktree created via `git worktree add`) `.git`
is a file that points to a worktree git dir (within the repo's `.git`).  This
directory contains `HEAD` and a file that contains the path to the repo's git
dir.

`findRepo` now returns a POJO with two paths: the worktree git dir and the
common git dir.  For the base work tree, these paths will be the same.  For a
linked worktree, they will be different, eg

```javascript
{
  worktreeGitDir: '<GIT_DIR>/worktrees/base',
  commonGitDir:   '<GIT_DIR>',
}
```

[Fix #25]